Remove GCed memory from MainLoop
Since MainLoop is created with `create` the sequence stored within it
isn't visible to the GC. So when it runs it doesn't find the reference
to it and can therefore end up freeing it. This commit fixes #17 which
was caused by such a scenario (probably because the Firefox notification
included an icon which contains enough data to have triggered the GC).
This also means that MainLoop can now be free'd manually if the user
wishes to use `create` and `tick` without `runForever`.